--- module_utils/ca_common.py
@@ -1,11 +1,0 @@
-def exec_command(module, cmd, stdin=None):
-    '''
-    Execute command(s)
-    '''
-
-    binary_data = False
-    if stdin:
-        binary_data = True
-    rc, out, err = module.run_command(cmd, data=stdin, binary_data=binary_data)
-
-    return rc, cmd, out, err
